requirements.txt
django==2.2.8
graphene-django==2.2.0
django-filter==2.0.0
django-graphql-jwt==0.1.5
django-model-utils==4.0.0
pycurl==7.43.0.3
djangorestframework==3.11.0
psycopg2==2.8.4
django-extensions==2.2.5
dj-database-url==0.5.0
mysqlclient==1.4.6
requests==2.22.0
django-dynamic-fixture==2.0.0
pytest==5.3.2
pytest-django==3.7.0
pytest-mock==1.13.0
pytest-pythonpath==0.7.3
